vdo: fix 32b portability issue in vdo_header struct
Replace size_t with uint64_t in struct vdo_header to ensure consistent
64-bit size field across all architectures.

On 32-bit platforms, size_t is only 4 bytes, which causes incorrect
structure layout when reading VDO metadata that expects a 64-bit size
field. This fix ensures the VDO header structure maintains the same
binary layout regardless of target architecture.

While 32-bit architectures are not officially supported for VDO,
this change improves code correctness and prevents potential issues
during cross-compilation or when building on 32-bit development
environments.

+++ device_mapper/vdo/vdo_reader.c
@@ -67,7 +67,7 @@ enum {
 struct vdo_header {
 	uint32_t id; /* The component this is a header for */
 	struct vdo_version_number version; /* The version of the data format */
-	size_t size; /* The size of the data following this header */
+	uint64_t size; /* The size of the data following this header */
 } __packed;
 
 struct vdo_geometry_block {
